requirements.txt
flask==2.3.2
flask-cors==3.0.10
google-generativeai==0.3.2
python-dotenv==1.0.0
black==24.4.2